agent/requirements.txt
google-adk
google-cloud-firestore
python-dotenv
fastapi
uvicorn
